refactor get_annotation
