+++ python/essex-config/essex_config/sources/source.py
@@ -3,7 +3,7 @@
 from abc import ABC, abstractmethod
 from typing import Any, TypeVar
 
-from essex_config.field_decorators import Alias, Parser
+from essex_config.field_annotations import Alias, Parser
 from essex_config.sources.convert_utils import convert_to_type
 
 SourceValueType = TypeVar("SourceValueType")
