+++ python/essex-config/tests/integration/graphrag_config/input_config.py
@@ -8,7 +8,7 @@ from typing import Annotated
 from pydantic import BaseModel, Field
 
 import tests.integration.graphrag_config.defaults as defs
-from essex_config.field_decorators import Parser
+from essex_config.field_annotations import Parser
 from essex_config.sources.utils import plain_text_list_parser
 
 from .enums import InputFileType, InputType
